+++ src/test/java/seedu/address/logic/parser/util/ArgumentTokenizerTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.util;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
